CycleGAN/pyproject.toml
[tool.poetry]
name = "cyclegan"
version = "0.2.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.3"
pillow = "^11.0.0"
tqdm = "^4.67.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
albumentations = "^1.4.21"
pathlib = "^1.0.1"
requests = "^2.32.3"
matplotlib = "^3.9.2"
ruff = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
